Client wireguard registration
